Remove the action toolbar.
