+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/Compose/Toolbar/ToolbarButton.cs
@@ -5,6 +5,8 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
+using osu.Framework.Input.Events;
+using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
 using osuTK;
 
@@ -15,33 +17,11 @@ namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Compose
 /// </summary>
 public abstract partial class ToolbarButton : OsuClickableContainer
 {
-    public void SetIcon(Drawable icon)
-    {
-        Size = new Vector2(SpecialActionToolbar.HEIGHT);
-        IconContainer.Icon = icon;
-        IconContainer.Show();
-    }
-
     [Resolved]
     private TextureStore textures { get; set; } = null!;
 
-    public void SetIcon(string texture) =>
-        SetIcon(new Sprite
-        {
-            Texture = textures.Get(texture),
-        });
-
-    public void SetIcon(IconUsage iconUsage) =>
-        SetIcon(new SpriteIcon
-        {
-            Icon = iconUsage,
-        });
-
-    protected void SetState(bool enabled)
-    {
-        IconContainer.Icon.Alpha = enabled ? 1f : 0.5f;
-        Enabled.Value = enabled;
-    }
+    [Resolved]
+    private OsuColour colours { get; set; } = null!;
 
     protected ConstrainedIconContainer IconContainer;
 
@@ -57,5 +37,43 @@ public abstract partial class ToolbarButton : OsuClickableContainer
                 Alpha = 0,
             },
         };
+    }
+
+    public void SetIcon(IconUsage iconUsage) =>
+        SetIcon(new SpriteIcon
+        {
+            Icon = iconUsage,
+        });
+
+    public void SetIcon(Drawable icon)
+    {
+        Size = new Vector2(SpecialActionToolbar.HEIGHT);
+        IconContainer.Icon = icon;
+        IconContainer.Show();
+    }
+
+    protected void SetState(bool enabled)
+    {
+        IconContainer.Icon.Alpha = enabled ? 1f : 0.5f;
+        Enabled.Value = enabled;
+    }
+
+    protected override bool OnClick(ClickEvent e)
+    {
+        ToggleClickEffect();
+
+        return base.OnClick(e);
+    }
+
+    public void ToggleClickEffect()
+    {
+        if (Enabled.Value)
+        {
+            IconContainer.FadeOut(100).Then().FadeIn();
+        }
+        else
+        {
+            IconContainer.FadeColour(colours.Red, 100).Then().FadeColour(Colour4.White);
+        }
     }
 }
